add responsive nav bar
