+++ front/src/components/Navigation/SideDrawer.css
@@ -6,5 +6,5 @@
   height: 100vh;
   width: 70%;
   background: white;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.26);
+  /* box-shadow: 0 2px 8px rgba(0, 0, 0, 0.26); */
 }
